Switch from 'thwbr' env to direct swath application
The 'thwbr' environment uses \verbwrite{} command to write text
to working files, which will be word-segmented by an external tool.
However, this does not work with 'utf8x' inputenc, as the written
file contents will be in macros like \thaiKoKai, etc., which cannot
be readily analyzed by word segmentation tools.

* {chapters}.tex:
  - Strip off 'thwbr' wrapping environments and '\wbrin' commands.
* {chapters}.tex -> {chapters.ltx}:
  - Rename source files to be input to word segmentation tools,
    to get *.tex as the analyzed outputs to be included in LaTeX doc.
* Makefile:
  - Rewrite to preprocess the listed *.ltx with word segmentation tool
    before compiling the main doc.
  - Directly generate PDF with 'pdflatex'.
  - Drop 'bibtex' rule due to missing BibTeX source file.